Fix passing s3Client to `monolambda/s3`
monolambda/s3 client accepts an options argument, which is an object and not an instance of s3 client. I found this issue when I was trying to use the plugin for syncing with another profile (non default). Because s3Client was not passed correctly it defaulted to the `AWS.S3` with default profile, which resulted in AccessDenied error.